feat: created custom components for password and textfields that integrate with react-hook-form
